LayoutTests/js/dfg-use-function-as-variable-expected.txt

Tests what happens when you use a function as a variable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS run_tests() is 26
PASS successfullyParsed is true

TEST COMPLETE
